Starting to move the interface equilibrium constraints
